requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.12.0
bitsandbytes==0.49.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.0
langchain-core==1.2.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
markupsafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
optimum==2.1.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pillow==12.0.0
psutil==7.2.0
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
sentencepiece==0.2.1
six==1.17.0
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
uuid-utils==0.12.0
xxhash==3.6.0
zstandard==0.25.0